Tulpa origin, hashed out controversies & IAA
Corrected that it's actually not much from Tibetan Buddism apart from terminology, added source for that, reworded paragraph about cultural appropation a bit, because it's hypocritical of me to speak about ditching the word tulpa like I belive that it's what we all should do when I don't do it myself. Commented out controversies, because I want this page to be more educational than argumentative and I plan to move them somewhere else. And clarified that IAA is a scientific name for soulbond phenomena. If you're reading this, hello, didn't expected anybody here, but have a nice day. Remember about your water and maybe go for a walk or something.