Sources on tulpas & removed paragraph
Reworded the origin of tulpamancy a bit (I'm still a bit confused about what Tibetan practices are actually about and how they differ from Western tulpomancy, but I've heard Jacob Isler wants to write an article on the topic, so I'm waiting for that). Removed whole "Polish tulpa community may not know about the whole plural community", since the Polish tulpa community is basically dead anyway and it's unlikely to encounter a tulpamancer that is not also in the English community. Added that some Tibetan Buddists don't mind Western tulpomancy community with sources on that and removed the stance about the need to have the cultural appropiation in mind.