Added Arousal non-concordance to resources
I've watched a video on this topic in context of OCD and regrettably scrolled down to comments, where some of the audience was calling people like me "sickos" and assuming all pedophiles are monsters that want to hurt others. Gods, no wonders they are so distressed by these thoughts if they are so disgusted by people who have them. I belive I've written this to cope. It helped a little and some people may need it, so I guess it's a win-win. Now excuse me, I need to somehow pull myself together and learn what I need to know for two tomorrow's classes.