Fixing mobile menu the right way
Actually I did fuck it up and had no idea. I only figured it out, because I conicidentally looked into settings today and noticed a function with the same name. Like my programming teacher always said, it's never the computer's fault... 
But it means it was like this for a month 😭